+++ ios/Classes/NotificationWhenAppIsKilledPlugin.swift
@@ -11,6 +11,8 @@ public class NotificationWhenAppIsKilledPlugin: NSObject, FlutterPlugin {
   private var observerAdded = false
   private var notificationTitleOnKill: String!
   private var notificationBodyOnKill: String!
+  private var notificationInterruptionLevel: Int!
+  private var notificationUseDefaultSound: Boolean!
 
   public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
     switch call.method {
@@ -19,6 +21,8 @@ public class NotificationWhenAppIsKilledPlugin: NSObject, FlutterPlugin {
 
         notificationTitleOnKill = args["title"] as! String
         notificationBodyOnKill = args["description"] as! String
+        notificationInterruptionLevel = args["argsForIos"]["interruptionLevel"] as! Int
+        notificationUseDefaultSound = args["argsForIos"]["useDefaultSound"] as! Boolean
 
         observerAdded = true
         NotificationCenter.default.addObserver(self, selector: #selector(applicationWillTerminate(_:)), name: UIApplication.willTerminateNotification, object: nil)
@@ -41,6 +45,14 @@ public class NotificationWhenAppIsKilledPlugin: NSObject, FlutterPlugin {
           content.body = notificationBodyOnKill
           let trigger = UNTimeIntervalNotificationTrigger(timeInterval: 1, repeats: false)
           let request = UNNotificationRequest(identifier: "notification on app kill", content: content, trigger: trigger)
+
+          if (@available(iOS 15.0, *)) {
+            content.interruptionLevel = notificationInterruptionLevel;
+          }
+
+          if (notificationUseDefaultSound) {
+            content.sound = UNNotificationSound.defaultSound;
+          }
 
           UNUserNotificationCenter.current().add(request) { (error) in
               if let error = error {